#20 move roof converter in the appropriate package, optimized the imports, re added check on telescope for opening the curtains
