+++ crac_server/service/roof_service.py
@@ -1,9 +1,11 @@
 import logging
-from crac_protobuf.roof_pb2_grpc import (
-    RoofServicer,
+from crac_protobuf.roof_pb2_grpc import RoofServicer
+from crac_server.converter.roof_converter import RoofMediator
+from crac_server.handler.roof_handler import (
+    RoofCurtainsHandler, 
+    RoofHandler, 
+    RoofTelescopeHandler,
 )
-from crac_server.handler.roof_converter import RoofMediator
-from crac_server.handler.roof_handler import RoofCurtainsHandler, RoofHandler, RoofTelescopeHandler
 
 
 logger = logging.getLogger(__name__)
